random bench refactor: move to new bench/random.cpp
